chore: handle other attribute types (#101)
* chore: handle other attribute types

* fix: move common entirety list text widget

* chore: change attribute type to charfield in devices app

* fix: error while merging

* chore: update filip version

* chore: disable type validator of get_attributes

---------
