+++ app/Entirety/entirety/widgets.py
@@ -15,3 +15,20 @@ class SelectTextInputWidget(forms.MultiWidget):
 
     def decompress(self, value):
         return ["", ""]
+
+
+class ListTextWidget(forms.TextInput):
+    def __init__(self, data_list, name, *args, **kwargs):
+        super(ListTextWidget, self).__init__(*args, **kwargs)
+        self._name = name
+        self._list = data_list
+        self.attrs.update({"list": "list__%s" % self._name, "autocomplete": "off"})
+
+    def render(self, name, value, attrs=None, renderer=None):
+        text_html = super(ListTextWidget, self).render(name, value, attrs=attrs)
+        data_list = '<datalist id="list__%s">' % self._name
+        for item in self._list:
+            data_list += '<option value="%s">' % item
+        data_list += "</datalist>"
+
+        return text_html + data_list
