+++ app/Entirety/entities/forms.py
@@ -2,24 +2,8 @@ from crispy_forms.helper import FormHelper
 from crispy_forms.layout import Layout, Div, HTML
 from django import forms
 
+from entirety.widgets import ListTextWidget
 from entities.requests import AttributeTypes, get_entities_types
-
-
-class ListTextWidget(forms.TextInput):
-    def __init__(self, data_list, name, *args, **kwargs):
-        super(ListTextWidget, self).__init__(*args, **kwargs)
-        self._name = name
-        self._list = data_list
-        self.attrs.update({"list": "list__%s" % self._name})
-
-    def render(self, name, value, attrs=None, renderer=None):
-        text_html = super(ListTextWidget, self).render(name, value, attrs=attrs)
-        data_list = '<datalist id="list__%s">' % self._name
-        for item in self._list:
-            data_list += '<option value="%s">' % item
-        data_list += "</datalist>"
-
-        return text_html + data_list
 
 
 class EntityForm(forms.Form):
@@ -68,8 +52,19 @@ class AttributeForm(forms.Form):
             }
         ),
     )
-    type = forms.ChoiceField(
-        label="Attribute Type", choices=[(x.value, x.name) for x in AttributeTypes]
+    type = forms.CharField(
+        required=True,
+        max_length=256,
+        label="Attribute Type",
+        widget=ListTextWidget(
+            data_list=[x.value for x in AttributeTypes],
+            name="attr-type-list",
+            attrs={
+                "data-bs-toggle": "tooltip",
+                "data-bs-placement": "top",
+                "title": "Type of the context attribute.",
+            },
+        ),
     )
     value = forms.CharField(
         required=False,
