+++ app/Entirety/requirements.txt
@@ -8,7 +8,7 @@ django_tables2==2.4.1
 django-pydantic-settings==0.6.3
 django-compressor~=4.1
 django-libsass==0.9
-filip~=0.2.3
+filip~=0.2.5
 mozilla_django_oidc==2.0.0
 pydantic==1.10.2
 crispy-bootstrap5==0.7
